Bug fix (Nan); better inversion

+++ real.hpp
@@ -1,6 +1,8 @@
 #ifndef REAL_H
 #define REAL_H
 
-typedef float real;
+// some .obj files contain doubles themselves, so using float
+// would corrupt the reader.
+typedef double real;
 
 #endif
